move typings to types directory; add testing file
